requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
ccxt==4.4.73
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
dateparser==1.2.1
Django==4.2.20
django-appconf==1.1.0
django-auditlog==3.0.0
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-cryptography==1.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
fastapi==0.115.12
firebase-admin==6.7.0
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.2
msgpack==1.1.0
multidict==6.3.2
mysqlclient==2.1.1
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.6.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-binance==1.0.28
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
razorpay==1.4.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.19.0
zope.interface==7.2
